+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added new placeholder to get amount of players in arena without spectator amount:
   * `%oitc_{arena_id}:players_left%`
 * Added enable option for spectator settings.
+* Fixed possible NPE when registering commands.
 * Fixed getting NPE when place signs another world.
 * Fixed scoreboard friendly fire caused players visible.
 * Fixed NPE when signs aren't located in default world.